Initial Reconfiguration of Types
