--- adminPrivileges.ts
@@ -1,4 +1,0 @@
-export enum AdminPrivileges {
-    MANAGER = 0,
-    SUPERUSER = 1,
-}
