+++ analytics.ts
@@ -1,9 +1,10 @@
-import { reportType, CategoryTypes } from "..";
-
 /**
- * Legacy code but need to verify
+ * Sherlock Analytics Types
  */
-export class WeeklyOverallStats {
+
+import { CategoryTypes, reportType } from "./reports";
+
+ export class WeeklyOverallStats {
     public totalWeekCount: number | undefined;
     public changeTotalWeek: string | undefined;
     public amountChangeTotalWeek: number | undefined;
@@ -14,10 +15,10 @@ export class WeeklyOverallStats {
 
     checkInfo(): boolean {
         return (
-            this.totalWeekCount != undefined &&
-            this.changeTotalWeek != undefined &&
-            this.amountChangeTotalWeek != undefined
-        )
+            this.totalWeekCount !== undefined &&
+            this.changeTotalWeek !== undefined &&
+            this.amountChangeTotalWeek !== undefined
+        );
     }
 }
 
@@ -25,86 +26,68 @@ export class MonthlyOverallStats {
     public totalMonthCount: number | undefined;
     public changeTotalMonth: string | undefined;
     public amountChangeTotalMonth: number | undefined;
-
     constructor(init?: Partial<MonthlyOverallStats>) {
         Object.assign(this, init);
     }
-
     checkInfo(): boolean {
         return (
-            this.totalMonthCount != undefined &&
-            this.changeTotalMonth != undefined &&
-            this.amountChangeTotalMonth != undefined
-        )
+            this.totalMonthCount !== undefined &&
+            this.changeTotalMonth !== undefined &&
+            this.amountChangeTotalMonth !== undefined
+        );
     }
 }
 
 export class StatsPerCategory {
-    public name: reportType | CategoryTypes | undefined;
+    public name: string | undefined;
     public count: number | undefined;
     public change: string | undefined;
     public amountChange: number | undefined;
-
     constructor(init?: Partial<StatsPerCategory>) {
         Object.assign(this, init);
     }
-
     checkInfo(): boolean {
         return (
-            this.name != undefined &&
-            this.name != undefined &&
-            this.count != undefined &&
-            this.change != undefined &&
-            this.amountChange != undefined
-        )
+            this.name !== undefined &&
+            this.name !== "" &&
+            this.count !== undefined &&
+            this.change !== undefined &&
+            this.amountChange !== undefined
+        );
     }
 }
 
-// We are using this in dataApplied in TimePicker so not sure if we can get rid on this or not
 export class ThirtyDayCount {
     public dateTime: string | undefined;
     public dayName: string | undefined;
     public count: number | undefined;
-
     constructor(init?: Partial<ThirtyDayCount>) {
         Object.assign(this, init);
     }
-
     checkInfo(): boolean {
-        return (
-            this.dateTime != undefined &&
-            this.dayName != undefined &&
-            this.count != undefined
-        )
+        return this.dateTime !== undefined && this.dayName !== undefined && this.count !== undefined;
     }
 }
-
 export class Analytics {
     public weeklyOverallStats: WeeklyOverallStats | undefined;
     public monthlyOverallStats: MonthlyOverallStats | undefined;
     public weeklyStatsPerCategory: StatsPerCategory[] | undefined;
     public monthlyStatsPerCategory: StatsPerCategory[] | undefined;
     public thirtyDayCount: ThirtyDayCount[] | undefined;
-
     constructor(init?: Partial<Analytics>) {
         Object.assign(this, init);
     }
-
     checkInfo(): boolean {
         return (
-            this.weeklyOverallStats != undefined &&
-            this.weeklyStatsPerCategory != undefined &&
-            this.monthlyOverallStats != undefined &&
-            this.monthlyStatsPerCategory != undefined &&
-            this.thirtyDayCount != undefined
-        )
+            this.weeklyOverallStats !== undefined &&
+            this.weeklyStatsPerCategory !== undefined &&
+            this.monthlyOverallStats !== undefined &&
+            this.monthlyStatsPerCategory !== undefined &&
+            this.thirtyDayCount !== undefined
+        );
     }
 }
 
-
-/**
- * Sherlock V2 analytics summary
- */
 export class PerCategoryStats {
     public category: reportType | CategoryTypes | undefined;
     public count: number | undefined;
@@ -151,7 +134,6 @@ export class CategoryStats {
         )
     }
 }
-
 
 export class AnalyticsSummary {
     public perCategoryStats: PerCategoryStats[] | undefined;
