--- login.ts
@@ -1,11 +1,0 @@
-export class Login {
-  public iat: string;
-  public exp: string;
-  public token: string;
-
-  constructor(iat: string, exp: string, token: string) {
-    this.iat = iat;
-    this.exp = exp;
-    this.token = token;
-  }
-}
