+++ status.ts
@@ -1,4 +1,1 @@
-export enum Status {
-  SUCCESS,
-  ERROR,
-}
+
